+++ BlazorSplitGrid/Models/Track.cs
@@ -1,9 +1,9 @@
 namespace BlazorSplitGrid.Models;
 
-public record Track(string Id, int Number, bool IsGutter, decimal Size, decimal InitialSize, decimal MinSize, decimal MaxSize, string Selector)
+public record Track(string Id, int Number, bool IsGutter, string Size, string InitialSize, int MinSize, int MaxSize, string Selector)
 {
     public string ToSizeString()
     {
-        return $"{Size}{(IsGutter ? "px" : "fr")}";
+        return $"{Size}";
     }
 }